services/indexing-service/requirements.txt
docling
requests
weaviate-client==3.26.0
python-dotenv
fastapi
uvicorn
python-multipart